% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{API_URL}
\alias{API_URL}
\title{COVID-19 api link: https://github.com/GregVes/covid-19-api-server}
\format{An object of class \code{character} of length 1.}
\usage{
API_URL
}
\description{
COVID-19 api link: https://github.com/GregVes/covid-19-api-server
}
\keyword{datasets}
