% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{covid_api_request}
\alias{covid_api_request}
\title{Sends request to COVID-19 API}
\usage{
covid_api_request(request)
}
\arguments{
\item{request}{character with request to api, eg. "?country=Spain" if you
want to display data for Spain}
}
\value{
list with response and parsed response from json
}
\description{
details on API syntax: \url{https://github.com/GregVes/covid-19-api-server}
}
\examples{
{ 
covid_api_request("?country=Spain")
}
}
