% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_response_status}
\alias{.check_response_status}
\title{Check response status}
\usage{
.check_response_status(resp, expected_code, error_message = "Request failed")
}
\arguments{
\item{resp}{httr response}

\item{expected_code}{numeric with expected code e.g. 201}

\item{error_message}{character with error message}
}
\description{
Verifies if response status is correct.
If not it stops execution with message.
}
